refactor(vtree): lots of adjustments, tweaks, fixes
