+++ packages/vtree/core/modeling/operations/measurements/measureBounds.js
@@ -1,5 +1,5 @@
-const cacheWithInvalidation = require('./cacheWithInvalidation')
-const cachedGenerator = require('./geometry-generator-cached')
+const cacheWithInvalidation = require('../../../cacheWithInvalidation')
+const cachedGenerator = require('../../../generators/geometry-generator-cached-csg')
 
 /**
  * Unlike most 'virtual' functions measurement of bounds pre-computes the needed geometric data
