+++ packages/vtree/core/modeling/operations/transforms/center.js
@@ -1,0 +1,8 @@
+const { toArray } = require('@jscad/array-utils')
+
+const center = (...solids) => {
+  solids = toArray(solids)
+  return { children: solids, type: 'center', params: undefined }
+}
+
+module.exports = center
